+++ test/test_constants.dart
@@ -8,3 +8,5 @@ MHcCAQEEIC7MfY+wECbErLMx+7w87cWuAs6tdm505p9yGfEaxqVfoAoGCCqGSM49
 AwEHoUQDQgAEoUiHOzjV7KODS686wHTYwY/gz5dvPKjVS/maWU+VKcuNBUvAc5hj
 hSHpKl+DhF2u09WahshD18zlBVqFGO46Fg==
 -----END EC PRIVATE KEY-----''';
+final String? skipTests = envVars['SKIP_TESTS'];
+final bool ciSkip = skipTests == 'false' ? false : true;