+++ lib/models.dart
@@ -1,4 +1,5 @@
 export 'src/models/account.dart';
+export 'src/models/candle.dart';
 export 'src/models/credential.dart';
 export 'src/models/fee_tier.dart';
 export 'src/models/fill.dart';
@@ -14,6 +15,7 @@ export 'src/models/order_configuration/stop_limit_gtd.dart';
 export 'src/models/portfolio.dart';
 export 'src/models/portfolio_breakdown.dart';
 export 'src/models/product.dart';
+export 'src/models/product_book.dart';
 export 'src/models/server_time.dart';
 export 'src/models/signature.dart';
 export 'src/models/trade.dart';